FIX: Commenting out the non deterministic session file garbage collection
the session_gc method runs in the main session process/thread and even with the randomness, some session will get into the cleanup part and if there are huge number session files, it will be a drag. Especially so since odoo creates anonymous session files and that adds up pretty quickly.
The out-of-band mitigation is to have a cron that will specifically target certain session files while removing all others that are not needed. session_redis is another option since redis will expire the session automatically.